requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.4.1
Async-OAuthlib==0.0.9
async-timeout==3.0.1
attrs==21.4.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
discord.py @ git+https://github.com/Rapptz/discord.py@fd9c244f480ae12f972daf2ffe34ae83125847fe
distlib==0.3.3
Django==3.2.9
dnspython==2.1.0
easy-pil==0.1.3
fake-useragent==0.1.11
filelock==3.3.2
Flask==2.0.1
h11==0.12.0
h2==4.0.0
heroku==0.1.4
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.8.1
inv==0.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonlines==2.0.0
key-generator==1.0.3
lxml==4.6.3
MarkupSafe==2.0.1
MouseInfo==0.1.3
multidict==6.0.2
mysql-connector-python==8.0.26
numpy==1.21.2
oauthlib==3.1.1
pafy==0.5.5
parse==1.19.0
phonenumbers==8.12.31
Pillow==8.4.0
platformdirs==2.4.0
praw==7.4.0
prawcore==2.3.0
priority==2.0.0
protobuf==3.19.3
PyAutoGUI==0.9.52
pycparser==2.20
pyee==8.2.2
pygame==2.0.1
PyGetWindow==0.0.9
PyJWT==2.1.0
pymongo==3.12.0
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyperclip==1.8.2
pyppeteer==0.2.6
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquery==1.4.3
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==1.5
python-dotenv==0.19.0
PyTweening==1.0.3
pytz==2021.1
Quart==0.15.1
Quart-Discord==2.1.4
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tabulate==0.8.9
toml==0.10.2
tqdm==4.62.3
typing==3.7.4.3
typing-extensions==4.3.0
tzlocal==2.1
update-checker==0.18.0
urllib3==1.26.4
virtualenv==20.10.0
w3lib==1.22.0
WebOb==1.8.7
websocket-client==1.2.1
websockets==9.1
Werkzeug==2.0.1
wsproto==1.0.0
yarl==1.7.2
youtube-dl==2021.6.6
zipp==3.6.0